+++ tests/integration/static_credentials_test.go
@@ -1,7 +1,7 @@
 //go:build !fast
 // +build !fast
 
-package tests
+package integration
 
 import (
 	"context"
@@ -20,8 +20,6 @@ import (
 )
 
 func TestStaticCredentials(t *testing.T) {
-	t.Skip("wait for newest cr.yandex/yc/yandex-docker-local-ydb:latest was published")
-
 	ctx, cancel := context.WithTimeout(context.Background(), time.Second*10)
 	defer cancel()